moved email functionality to common utilities package
